+++ packages/stellar-wallet-snap/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `signProofOfOwnership` client request for silent proof-of-ownership signing (SEP-0053)
+
 ## [0.1.0]
 
 ### Added